Shorten default session
I was still getting timeouts and connections really don't need to be
retained that long do they?

Also added an explicit commit on the price updater because for some
reason I didn't want to debug it wasn't committing the transaction at
the end of the session. Due to that change I switched the response
code to the POST to a more conventional 200.

Also change logger to the general best-practice for logging.